Combined location list creation into a higher level
