+++ docs/Roadmap.md
@@ -5,3 +5,8 @@
 - Implants not factored in to job duration
 - Tax rates should be set individually for each activity
 
+
+
+# Other cleanup
+
+- Add "Eve" namespace to all Eve-related data types
